added the report user window
- refactor the hooks in video window's index to separate file
